Page Structure, grey theme, splash screen, loading imrovments
