+++ src/App.js
@@ -3,6 +3,7 @@ import { Route, Switch } from 'react-router-dom';
 import './App.css';
 import SignIn from './components/SignIn';
 import SongDetail from './components/SongDetail';
+import LoadingCloud from './components/LoadingCloud';
 import Search from './connected/Search';
 import Navbar from './connected/Navbar';
 import ArtistPage from './connected/ArtistPage';
@@ -10,10 +11,11 @@ import { setUser } from './redux/actions';
 import { Redirect } from 'react-router-dom';
 import * as selectors from './redux/selectors';
 import { connect } from 'react-redux';
-import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
+import { Paper } from '@material-ui/core';
 import paths from './paths.js';
 import ReactGA from 'react-ga';
 import { history } from './redux/store';
+import SplashScreen from './components/SplashScreen';
 
 function initializeReactGA() {
 	console.log('Initializing analytics');
@@ -25,25 +27,6 @@ function initializeReactGA() {
 	});
 }
 
-const theme = createMuiTheme({
-	background: 'linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)',
-	palette: {
-		primary: {
-			light: '#6d6d6d',
-			main: '#424242',
-			dark: '#1b1b1b',
-			contrastText: '#ffffff'
-		},
-		secondary: {
-			light: '#64c1ff',
-			main: '#0091ea',
-			dark: '#0064b7',
-			contrastText: '#f5f5f5'
-		}
-	},
-	type: 'dark'
-});
-
 const App = (props) => {
 	const { user, appIsHydrated, location } = props;
 	useEffect(() => {
@@ -53,37 +36,36 @@ const App = (props) => {
 		console.log('APP rendered w/ no user after hydration, redirecting.');
 		return <Redirect to={paths.signIn} />;
 	}
-	return (
-		<ThemeProvider theme={theme}>
-			{appIsHydrated ? (
-				<div className="App">
-					<Navbar />
-					<Switch>
-						<Route path={paths.signIn} render={(routerProps) => <SignIn history={routerProps.history} />} />
-						<Route path={paths.search} render={(routerProps) => <Search history={routerProps.history} />} />
-						<Route
-							path={paths.songPage}
-							render={(routerProps) => <SongDetail history={routerProps.history} />}
-						/>
-						<Route
-							path={paths.artistPage}
-							render={({ history }) => {
-								return <ArtistPage history={history} />;
-							}}
-						/>
-						<Route render={() => <Redirect to={paths.search} />} />
-					</Switch>
-				</div>
-			) : (
-				<h1>Rap Clouds</h1>
-			)}
-		</ThemeProvider>
+	return appIsHydrated ? (
+		<Paper style={{ height: '100vh', width: '100vw', overflow: 'hidden' }}>
+			<Navbar />
+			<Switch>
+				<Route path={paths.signIn} render={(routerProps) => <SignIn history={routerProps.history} />} />
+				<Route path={paths.search} render={(routerProps) => <Search history={routerProps.history} />} />
+				<Route path={paths.songPage} render={(routerProps) => <SongDetail history={routerProps.history} />} />
+				<Route
+					path={paths.artistPage}
+					render={({ history }) => {
+						return <ArtistPage history={history} />;
+					}}
+				/>
+				<Route render={() => <Redirect to={paths.search} />} />
+				{process.env.NODE_ENV === 'development' ? (
+					<React.Fragment>
+						<Route path={'/loadingCloud'} render={(routerProps) => <LoadingCloud />} />
+						<Route path={'/splash'} render={(routerProps) => <SplashScreen />} />
+					</React.Fragment>
+				) : null}
+			</Switch>
+		</Paper>
+	) : (
+		<h1>Rap Clouds</h1>
 	);
 };
 
 const mapState = (state) => ({
 	user: selectors.getUser(state),
-	appIsHydrated: selectors.isAppRehydrated(state)
+	appIsHydrated: selectors.isAppRehydrated(state),
 });
 
 export default connect(mapState, { setUser })(App);
