+++ src/index.js
@@ -9,8 +9,11 @@ import { Switch, Route, Link } from 'react-router-dom';
 import { ConnectedRouter } from 'connected-react-router';
 import { PersistGate } from 'redux-persist/es/integration/react';
 import { history } from './redux/store';
-import LoadingCloud from './components/LoadingCloud';
+import SplashScreen from './components/SplashScreen';
 import ErrorBoundary from './components/ErrorBoundary';
+import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
+import { Paper } from '@material-ui/core';
+
 if (process.env.NODE_ENV === 'development') {
 	window.store = store;
 }
@@ -20,33 +23,50 @@ const onBeforeLift = () => {
 	return;
 };
 
+const theme = createMuiTheme({
+	background: 'linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)',
+	palette: {
+		primary: {
+			light: '#6d6d6d',
+			main: '#424242',
+			dark: '#1b1b1b',
+			contrastText: '#ffffff'
+		},
+		secondary: {
+			light: '#64c1ff',
+			main: '#0091ea',
+			dark: '#0064b7',
+			contrastText: '#f5f5f5'
+		},
+		background: {
+			paper: '#6d6d6d',
+			default: '#424242'
+		}
+	},
+	type: 'dark'
+});
+
 ReactDOM.render(
 	<React.StrictMode>
 		<Provider store={store}>
-			<PersistGate
-				loading={
-					<div style={{ height: '100vh', width: '100vh', paddingTop: '40%' }}>
-						<div style={{ margin: 'auto' }}>
-							<LoadingCloud />
-						</div>
-					</div>
-				}
-				onBeforeLift={onBeforeLift}
-				persistor={persistor}
-			>
-				<ConnectedRouter history={history}>
-					<Switch>
-						<Route
-							path="/"
-							render={(routerProps) => (
-								<ErrorBoundary>
-									<App {...routerProps} />
-								</ErrorBoundary>
-							)}
-						/>
-					</Switch>
-				</ConnectedRouter>
-			</PersistGate>
+			<ThemeProvider theme={theme}>
+				<Paper style={{ height: '100vh', width: '100vw', overflow: 'hidden' }}>
+					<PersistGate loading={<SplashScreen />} onBeforeLift={onBeforeLift} persistor={persistor}>
+						<ConnectedRouter history={history}>
+							<Switch>
+								<Route
+									path="/"
+									render={(routerProps) => (
+										<ErrorBoundary>
+											<App {...routerProps} />
+										</ErrorBoundary>
+									)}
+								/>
+							</Switch>
+						</ConnectedRouter>
+					</PersistGate>
+				</Paper>
+			</ThemeProvider>
 		</Provider>
 	</React.StrictMode>,
 	document.getElementById('root')
